refactor: move RichTextEditor in own folder
In preparation for when it will be moved out of ER app.
