+++ src/components/DetailsPanel/InterpretationsUnit/RichTextEditor/RichTextEditor.js
