+++ src/components/DetailsPanel/InterpretationsUnit/RichTextEditor/styles/RichTextEditor.module.css
